% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C_to_F.R
\name{C_to_F}
\alias{C_to_F}
\title{Conversión de la temperatura en Centigrados a Fahrenheit}
\usage{
C_to_F(temp_centigrado)
}
\arguments{
\item{temp_centigrado}{un vector numérico de la temperatura en Centigrados}
}
\value{
Un vector numérico de la temperatura en grados fahrenheit.
}
\description{
La función \code{C_to_F()} toma un valor numérico de la temperatura en Centigrados y calcula la equivalencia en grados fahrenheit.
}
\details{
Estos son detalles extras de la función.
}
\examples{
C_to_F(23)

}
