% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{datos}
\alias{datos}
\title{Conjunto de datos de ejemplo}
\format{
Los datos son generados de manera aleatoria.
\describe{
\item{a}{muestra de datos}
\item{b}{muestra de datos multiplicada por 5}
}
}
\source{
Fuente: generación propia con funciones de R.
}
\usage{
datos
}
\description{
Los datos son utilizados para poder chequear nuestra función suma().
}
\keyword{datasets}
