% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/descargar_datos.R
\docType{data}
\name{descargar_datos}
\alias{descargar_datos}
\title{Datos de estaciones meteorológicas}
\format{
Un dataframe con las siguientes estaciones y columnas:
\describe{
\item{NH0098}{Datos de la estación meteorológica "NH0098".}
\item{NH0437}{Datos de la estación meteorológica "NH0437".}
\item{NH0046}{Datos de la estación meteorológica "NH0046".}
\item{NH0910}{Datos de la estación meteorológica "NH0910".}
\item{NH0472}{Datos de la estación meteorológica "NH0472".}
\item{NH_unidos}{Datos combinados de todas las estaciones.}
}
}
\source{
Repositorio de GitHub.
}
\usage{
descargar_datos

descargar_datos(id_estacion)
}
\arguments{
\item{id_estacion}{un string conformado por numeros y letras}
}
\value{
Devuelve la informacion contenida en el dataset descargado
}
\description{
Este conjunto de datos incluye muestras de diferentes estaciones meteorológicas, obtenidas de un repositorio en GitHub.

La funcion \code{descargar_datos()} toma 2 parametros y permite la lectura y/o
descarga del dataset de la estacion que se solicite.
}
\details{
Estaciones Disponibles:
\itemize{
\item NH0437
\item NH0098
\item NH0046
\item NH0910
\item NH0472
}
}
\examples{
data <- descargar_datos("NH0098")
}
\keyword{datasets}
