% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_to_C.R
\name{F_to_C}
\alias{F_to_C}
\title{Conversión de la temperatura en Fahrenheit a Centigrados}
\usage{
F_to_C(temp_fahrenheit)
}
\arguments{
\item{temp_fahrenheit}{un vector numérico de la temperatura en Fahrenheit}
}
\value{
Un vector numérico de la temperatura en grados centígrados.
}
\description{
La función \code{F_to_C()} toma un valor numérico de la temperatura en Fahrenheit y calcula la equivalencia en grados centígrados.
}
\details{
Estos son detalles extras de la función.
}
\examples{
F_to_C(100)

}
