% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suma.R
\name{suma}
\alias{suma}
\title{Suma de dos numeros positivos}
\usage{
suma(x, y)
}
\arguments{
\item{x}{un vector numérico positivo}

\item{y}{un vector numérico positivo}
}
\value{
Un vector numérico con la suma de \code{x} e \code{y}.
}
\description{
La función \code{suma()} dos valores numéricos positivos y calcula la suma entre ambos números.
}
\details{
Estos son detalles extras de la función.
}
\examples{
suma(10,10)
suma(5,14)

}
