% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabla_resumen.R
\name{tabla_resumen_temperatura}
\alias{tabla_resumen_temperatura}
\title{Tabla de resumen para la/s estacion/es meteorologicas}
\usage{
tabla_resumen_temperatura(data, estaciones)
}
\arguments{
\item{data}{un dataframe que contiene la informacion de las estaciones meteorologicas}

\item{estaciones}{id correspondiente a la/s estacion/es que deseamos investigar}
}
\value{
una tabla con el resumen de ciertas variables
}
\description{
Tabla de resumen para la/s estacion/es meteorologicas
}
\examples{
# Supomemos que queremos obtener la tabla de resumen de NH0910 y NH0437:
tabla_resumen_temperatura(NH_unidos,c('NH0910','NH0437'))

}
