% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descargar_datos.R
\name{descargar_datos}
\alias{descargar_datos}
\title{Descarga de datos de distintas estaciones}
\usage{
descargar_datos(id_estacion, ruta)
}
\arguments{
\item{id_estacion}{un string conformado por numeros y letras}

\item{ruta}{ubicación del dataset en nuestro proyecto}
}
\value{
Devuelve la informacion contenida en el dataset descargado
}
\description{
La funcion \code{descargar_datos()} toma 2 parametros y permite la lectura y/o
descarga del dataset de la estacion que se solicite.
}
\details{
Estaciones Disponibles:
\itemize{
\item NH0437
\item NH0098
\item NH0046
\item NH0910
\item NH0472
}
}
\examples{
data <- descargar_datos("NH0098", "data_raw/NH0098.csv")
}
