% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grafico_temp_mensual.R
\name{grafico_temperatura_mensual}
\alias{grafico_temperatura_mensual}
\title{Grafico del promedio mensual de la temperatura de abrigo}
\usage{
grafico_temperatura_mensual(
  datos,
  colores = c("blue", "red", "green"),
  titulo = "Temperatura"
)
}
\arguments{
\item{datos}{el data frame que contiene la informacion de todas las estaciones meteorologicas.}

\item{colores}{un vector de colores para utilizar en el grafico.}

\item{titulo}{un string con el titulo del grafico. Si no se define, el titulo por defecto sera "Temperatura".}
}
\value{
Un grafico con el promedio mensual de la temperatura.
}
\description{
Esta funcion recibe un data frame con datos de una o mas estaciones y
devuelve un grafico con el promedio mensual de la temperatura de abrigo
a 150 cm.
}
\examples{
grafico_temperatura_mensual(NH_unidos)
}
