% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grafico_temp_mensual.R
\name{grafico_temperatura_mensual}
\alias{grafico_temperatura_mensual}
\title{Gráfico del promedio mensual de la temperatura de abrigo}
\usage{
grafico_temperatura_mensual(
  datos,
  colores = c("blue", "red", "green"),
  titulo = "Temperatura"
)
}
\arguments{
\item{datos}{el data frame que contiene la información de todas las estaciones meteorológicas.}

\item{colores}{un vector de colores para utilizar en el gráfico.}

\item{titulo}{un string con el título del gráfico. Si no se define, el título por defecto será "Temperatura".}
}
\value{
Un gráfico con el promedio mensual de la temperatura.
}
\description{
Esta función recibe un data frame con datos de una o más estaciones y
devuelve un gráfico con el promedio mensual de la temperatura de abrigo
a 150 cm.
}
\examples{
grafico_temperatura_mensual(NH_unidos)
}
